Fix: Adding the TODO comment on b/437817546 for Abhinov branch.
Fix: Change name logging_mtc.py to validate_util.py. Since validate_util as the filename make more sense, we are validating the logs.

Fix: checkpoint_steps --> to Optional. Since this value exist but is not use in this test.

Fix: Change name mx-sv --> max-sv-loc. Helps with clarity in bucket checkpoints.

Fixes for code review
